Rework pagination to use fewer KV queries
Because KV currently does not provide a bulk read API, we have to query for each
artifact individually, which is expensive and gets more expensive as the page
size grows.

This commit reworks paging so that we instead store every artifact in the
archive in a single JSON array in a single KV key, and then we slice into that
to implement paging.

The tradeoff here is that now larger page sizes will be more efficient, because
each request to the Ace Archive API will require re-requesting the monolithic
artifact array from KV.

The hope is that this tradeoff will be worth it for two reasons:

- The number of artifacts in Ace Archive is quite small and will likely not grow
  very quickly. Even 1,000 artifacts, as a rough estimate, would likely not
  exceed 1-3 MiB in uncompressed payload size.
- Theoretically, the worker should have lower latency and higher bandwidth
  access to KV than the end user of the API. The hope is that implementing
  paging like this will be more efficient for end users than just serving them
  the monolithic artifact array from KV directly.